Move the UnfinishedCoGroupOperation class into its own Scala file.
The UnfinishedCoGroupOperation does not relate closely to CoGroupOperation
via sealed modifier so per Scala style guide [1] I propose to move it to
separate file.

[1] http://docs.scala-lang.org/style/files.html
