+++ flink-scala/src/main/scala/org/apache/flink/api/scala/CoGroupDataSet.scala
@@ -26,16 +26,13 @@ import org.apache.flink.api.common.InvalidProgramException
 import org.apache.flink.api.common.functions.{RichCoGroupFunction, CoGroupFunction}
 import org.apache.flink.api.common.functions.Partitioner
 import org.apache.flink.api.common.operators.Order
-import org.apache.flink.api.common.typeutils.TypeSerializer
 import org.apache.flink.api.java.operators._
-import org.apache.flink.api.java.typeutils.ObjectArrayTypeInfo
-import org.apache.flink.api.scala.typeutils.{CaseClassSerializer, CaseClassTypeInfo}
+import org.apache.flink.api.java.operators.Keys.ExpressionKeys
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.util.Collector
 import scala.collection.mutable
 import scala.collection.JavaConverters._
 import scala.reflect.ClassTag
-import org.apache.flink.api.java.operators.Keys.ExpressionKeys
 
 /**
  * A specific [[DataSet]] that results from a `coGroup` operation. The result of a default coGroup
@@ -273,69 +270,5 @@ class CoGroupDataSet[L, R](
       
       result
     }
-  }
-}
-
-/**
- * An unfinished coGroup operation that results from [[DataSet.coGroup]] The keys for the left and
- * right side must be specified using first `where` and then `isEqualTo`. For example:
- *
- * {{{
- *   val left = ...
- *   val right = ...
- *   val coGroupResult = left.coGroup(right).where(...).isEqualTo(...)
- * }}}
- * @tparam L The type of the left input of the coGroup.
- * @tparam R The type of the right input of the coGroup.
- */
-class UnfinishedCoGroupOperation[L: ClassTag, R: ClassTag](
-    leftInput: DataSet[L],
-    rightInput: DataSet[R])
-  extends UnfinishedKeyPairOperation[L, R, CoGroupDataSet[L, R]](leftInput, rightInput) {
-
-  private[flink] def finish(leftKey: Keys[L], rightKey: Keys[R]) = {
-    val coGrouper = new CoGroupFunction[L, R, (Array[L], Array[R])] {
-      def coGroup(
-                   left: java.lang.Iterable[L],
-                   right: java.lang.Iterable[R],
-                   out: Collector[(Array[L], Array[R])]) = {
-        val leftResult = Array[Any](left.asScala.toSeq: _*).asInstanceOf[Array[L]]
-        val rightResult = Array[Any](right.asScala.toSeq: _*).asInstanceOf[Array[R]]
-
-        out.collect((leftResult, rightResult))
-      }
-    }
-
-    // We have to use this hack, for some reason classOf[Array[T]] does not work.
-    // Maybe because ObjectArrayTypeInfo does not accept the Scala Array as an array class.
-    val leftArrayType =
-      ObjectArrayTypeInfo.getInfoFor(new Array[L](0).getClass, leftInput.getType)
-    val rightArrayType =
-      ObjectArrayTypeInfo.getInfoFor(new Array[R](0).getClass, rightInput.getType)
-
-    val returnType = new CaseClassTypeInfo[(Array[L], Array[R])](
-      classOf[(Array[L], Array[R])], Seq(leftArrayType, rightArrayType), Array("_1", "_2")) {
-
-      override def createSerializer: TypeSerializer[(Array[L], Array[R])] = {
-        val fieldSerializers: Array[TypeSerializer[_]] = new Array[TypeSerializer[_]](getArity)
-        for (i <- 0 until getArity) {
-          fieldSerializers(i) = types(i).createSerializer
-        }
-
-        new CaseClassSerializer[(Array[L], Array[R])](
-          classOf[(Array[L], Array[R])],
-          fieldSerializers) {
-          override def createInstance(fields: Array[AnyRef]) = {
-            (fields(0).asInstanceOf[Array[L]], fields(1).asInstanceOf[Array[R]])
-          }
-        }
-      }
-    }
-    val coGroupOperator = new CoGroupOperator[L, R, (Array[L], Array[R])](
-      leftInput.javaSet, rightInput.javaSet, leftKey, rightKey, coGrouper, returnType,
-      null, // partitioner
-      getCallLocationName())
-
-    new CoGroupDataSet(coGroupOperator, leftInput, rightInput, leftKey, rightKey)
   }
 }
